new component for description
